[HOOZ-103] FIX Admin login error layout bug (#60)
* [HOOZ-103] FIX Admin login error layout bug

When a mistake is made on the admin login form the validation errors
will be shown within the admin login form. The only consistent way to
achieve this was to set a session variable when the "/admin" path
was hit. Without using the session to store this state, when using
the forgotten password link the layout was incorrect and the login
button always redirected to the non-admin login.

- Use the session state for the default devise layouts in the initializer
- Set the session state for the admin layout on the sessions controller