[HOOZ-106] Division Developments CRUD (#61)
* [HOOZ-106] Division Developments CRUD

Developments CRUD has already been implemented under Developers, this PR
makes Developments CRUD available under Divisions as well.

A development can belongs_to either a developer or division, and a
development has many associations on Unit Type, Phase, Plot all
requiring a CRUD interface to work under each tree of breadcrumbs. To
make that work the child associations of development will need to use
Shallow Routes. This means that the breadcrumbs can generate the right
tree based on the parent of the development, and the url / view/
controller structure for unit types, phases and plots are not doing to
be repeated for developers and divisions.

- Ensure developments only have a single parent association (developer
  or division)
- CRUD views for developments under Divisions
- Sorting / Pagination for Division Developments
- Shallow routes on phases
- Breadcrumb updates to work with the Developer / Division tree

* Replace OwnableBy with InheritParentPermissionIds

On the developer's
divisions index table the division developments for that developer
should not be shown, so now Division developments require a blank
developer_id. This was previously required by the ownable_by macro to
assign the necessary foreign keys to the resources in the organization
hierarchy.

With the opportunity of this change, the ownable_by macro has been
dropped in favour or a model concern that is being used across resources
under developments. The development does not use the macro or concern.

- Create the InheritParentPermissionIds model concern
- Create the InheritPolymorphicPermissionIds model concern for documents
  and later faqs / contacts
- Use shared examples to encapsulate the specs around the model concern
- Remove lib/ownable_by.rb