Skip accessible_by for options when CF Admin (#112)
Choosing a developer with 1000s of developments associated with it crashes chrome on the
notifications and users forms. This is due to performance issues.

This commit introduces a few performance improvements for loading the cascading dropdowns:

- skip accessible_by calls if current user is a CF Admin
- Do not load options for the dropdowns when loading the view, leave that for AJAX call
- Only select used columns for building JSON responses

This has significantly reduced the response times on these pages, but the existing bottleneck
is building the JSON to return to populate each dropdown. Caching maybe a valid option
for this data later down the line.

The cascading dropdown inputs have all been refactored into a single helper method too :tada: