[HOOZ-88] Resident Permissions (#128)
To prepare for hooking up the My Home views, this PR sets up the CanCan abilities for residents. With this in place, displaying the correct set of appliances, rooms, documents, etc... should be as simple as going through the permissions `Room.accessible_by(current_resident)`.

The relationship between plots and residents has changed to a `has_one`, but still `through: :plot_residency`. The `PlotResidency` is where the completion date for a plot will live in future. This change supports the current domain model, requirements and known permissions. It would be easy to change this to a `has_many` relationship, but if more residents are needed I would question their role and permissions and look to capture those in separate model if applicable. We are in a flexible position on this topic.

The models that are covered for resident abilities are now:

- **Developer:** through foreign_key on plot
- **Division:** through foreign_key on plot
- **Development:** through foreign_key on plot
- **Phase:** through foreign_key on plot
- **Plot:** through residency
- **UnitType:** through plots unit_type_id foreign_key
- **Finish:** through rooms with the plots unit_type_id
- **Appliance:** through rooms with the plots unit_type_id
- **Document:** through developer and development ids on the plot
- **Faq:** through developer and development ids on the plot
- **Contact:** through developer and development ids on the plot
- **Notification:** through developer and development ids on the plot
- **Brand:** as a read only polymorphic association for each model in the hierarchy


NOTE: The polymorphic models are not checking for `division_id` foreign keys because each polymorphic model associated with a division automatically picks up the `developer_id`, via the included `PolymorphicPermissionable` concern.

---

Other changes:

- The tab helper can build `has_one` or `has_many` associations to check permissions against without needing to override the `permissions_on` option
- The `PolymorphicPermissionable` concern has been extended to set permission foreign keys when a polymorphic model is associated with a plot
- Permission foreign keys have been added to the `Faq` and `Contact` models, which now include the `PolymorphicPermissionable` concern to auto-inherit developer, division and develoment ids respectively
- A residents brand is now found using the permissions hierarchy to find the first match from `phase` upto `developer`
- The unused `Image` model, from the prototype days, has been exterminated
- A shared example spec has been created and used for the polymorphic read only access via a residency, to power the specs behind a residents `Document`, `Faq`, `Contact`, and `Notification` permissions