[HOOZ-44] Assign a resident to a plot (#134)
This commit uses the existing CRUD conventions used in nearly all other resource sections,
but to make this straight forward in the controller and views the PlotResidency model is
handling a lot of the complexity. The form itself contains fields to create or update an
existing resident, the completion date which lives on the PlotResidency, and is scoped
under the plot the user navigated to.

When creating a residency, either a new resident is created or an existing resident is updated,
editing the residency means the resident can also be updated. A plot can only have one
resident, of which cancan abilities have been setup to enforce this (even for CF Admins).

The cucumber feature only uses the development admin, instead of having scenarios for each
type of admin, because the other admins inherit the development admin permissions and
would not require any further code changes to be made.