[HOOZ-176] Homeowners library (#146)
When a homeowner visits their library page they should see a filtered set of the documents related to their plots,
choosing other categories should filter the documents by that category, **but** if the appliances categories is
chosen then all of the manuals related to appliances in their plot should be displayed.

When a PDF is uploaded the uploader will now create a preview image to show as a thumbnail in the library view.
The carrierwave_backgrounder gem has been installed to make the processing a background job, the strategy used
will upload the original document to S3 and then the background job will do the processing to create the preview image 
This is especially critical for the bulk documents upload function. The same background processing has been applied
to the appliance manuals and images.

To generate the previews of the thumbnails for existing documents there are rake tasks for the relevant models,
and a helper task that runs all `recreate_versions!` on all of the models in one go. The `tmp/uploads/` file requires
a permissions change to allow the rake task to put tmp files for processing in, so an .ebextensions job has been
added for that purpose.

The support files that are used in the cucumber suite have been converted to "valid" pngs, by stripping invalid chars,
which graphicsmagick output warnings for when running the tests. No other changes have been made to these files.

Other changes:

- Homeowner base controller now sets the `@plot` and `@brand`
- The brand template has been moved into the homeowner layout
- The preview_input helper now shows an upload button after a failed form submission which includes a new image,
  because we are behind a load balancer we do not have guaranteed access to the same tmp files and this can be
  misleading when re-submissions loose the new image.