[HOOZ-348] Custom plot rooms (#186)
CF Admins need the ability to customise a room for an individual plot. Rooms for plots are pulled from the plots unit type by default.

The new plot rooms controller allows rooms to be created that are attached to a plot, instead of a unit type, and can retain a link to the unit type room that was used as a template for the plot rooms attributes. With this data the `plot.rooms` method can exclude unit type rooms if any plot room have a link to them, creating a default template pattern.