HOTFIX: Add sitemap.xml to stop DDoS from Google
Search engines are using the sitemap from the old version of hoozzi.com, where
the how to articles are public content. Periodically, the crawlsers are bringing
down hoozzi.com becuase they are hitting urls that no longer exist and the
application is cannot handle that many exceptions being generated.

Files are not served from the public rails folder by default on elastic
beanstalk and there is not enough time to configure this, so instead there is a
rails route and a controller action that serves the content of the `sitemap.xml`.