Change to delayed job (#425)
The elasticache/redis/sidekiq combo was intermittently making the whole Hoozzi site unavailable. We spent some time investigating and failed to find the underlying cause, so instead have decided to move away from redis/sidekiq and use delayed_job instead.

This changes the architecture: async jobs will be stored in a database instead of in redis memory cache. There may be a small impact on performance, the database will now be shared between the async jobs and the primary runtime data.

We're not expecting this impact to be significant, there are not huge numbers of async jobs (sometimes the network traffic associated is large, but that should not be relevant to the DB impact)

There are two primary users of async jobs:

 *   uploading files / documents / images to S3 via carrierwave
 *   sending email notifications

We believe neither of these generates huge numbers of jobs

Also fixed
* alt text for document preview image