HOOZ-639 Fix apostrophes in admin and homeowner dashboard (#487)
The auto-generated FAQs have &rsquo in them. This doesn't cause any problems immediately.
Also, if you create an FAQ with ', that's fine too

But if you edit an auto-generated FAQ, and save it without fixing the apostrophe, it would show the &rsquo. This fixes the display of the title in both admin and homeowner dashboards.

I haven't fixed the edit page, which will still show the &rsquo, but you have the option of fixing it by hand when you do the edit, and even if you don't fix it, the dashboard displays are fixed.