Hooz 247 plot specific resident permissions (#585)
Move resident abilities from "invited by" column on resident to a "role" column on plot residency

This means we can capture different permissions for the same resident in the context of different plots, and specifically caters for the scenario where a resident is a homeowner on one plot, and a tenant on another. Plots may be in the same or different developers / divisions / developments

A homeowner can delete another resident if

 * the homeowner has resident permissions AND
 * the resident being deleted has either tenant or a homeowner permissions, and was added by a homeowner

A homeowner can not delete another resident if

 *  the resident was added by an admin OR
 *  the homeowner has tenant permissions

I've added the new role information to the development CSV, which records each resident separately.

The invited_by column is kept on DB and in development CSV, and is widely used in emails to new residents. It is no longer used for (most) permission testing. The exception is the delete permissions outlined above: invited_by doesn't change the ability of the resident doing the deleting, but it does affect the resident being deleted.

The invited_by column is removed from the residents list for a plot, and has been replaced by the role

A migration ports the existing plot invited_by onto existing plot_residencies: if there are multiple plot residencies for the same resident, or if there is no invited_by (which should predate the ability of residents to invite other residents), it defaults to "resident" role