Delete tenant (#598)
Defect fix: deleting a tenant would fail if the tenant was also associated with another plot.

To fix, I've moved the invited_by relationship across to the plot_residency table, and we reference this before deleting.

The old invited_by relationship has been left for the moment, in case there's an error during migration and we need to recover the old data.

Once the migration has completed successfully, the invited_by relationship can be removed from the resident table

Related spec tests

The JS changes are smaller than they look because I renamed a file