WeTransfer might be sending (#599)
Still prototyping, weTransfer file upload might be working, want to capture this quickly - but it isn't yet the final version. As before, I can not test weTransfer on local environment, only on Staging.

The URL is not correct, and the file name and other parameters may still change: but the file should be transferred successfully with this build.

Still no delete of the originals yet.