Upgrade cucumber (#633)
QA builds were failing with this codeship error:
Error: NOTE: Gem.gunzip is deprecated; use Gem::Util.gunzip instead. It will be removed on or after 2018-12-01.

The primary cultprit was cucumber, but updating cucumber had an impact on some other library dependencies, mainly relating to nokogiri. A full bundle update would affect the dependencies for the whole product, so instead I've explicilty updated the affected test libraries

The updated test libraries are much stricter on white space, so there are a bunch of which space fixes in the test suite as a result