HOOZ-248 (#696)
Working on the comments for HOOZ-248:

Tabs that are not populated are not visible.

Previously: Default home tab was set, so if the tab was not populated it would still load.
Now: The left-most populated tab is loaded.

Works for the sub_nav bar. Does not work for documents or contacts, as these are set up differently and I can't figure out how to recreate the above implementation in the context of either of these.